WIP: Add atan2 intrinsic function for HLSL
This commit adds support for the atan2 intrinsic function in HLSL. The function is implemented in the CGHLSLRuntime.h file and the IntrinsicsSPIRV.td file. It also adds the necessary changes in the Builtins.td, SemaHLSL.cpp, CGBuiltin.cpp, and hlsl_intrinsics.h files to handle the atan2 function. Additionally, test coverage for the atan2 function is added in the llvm/test/CodeGen/SPIRV/hlsl-intrinsics/atan2.ll file.